Lua updated to 5.3.5, sol 3.2.1 added

+++ DEM/Game/src/UI/PropUIControlSI.cpp
@@ -60,7 +60,7 @@ int CPropUIControl_AddActionHandler(lua_State* l)
 			CStrID(lua_tostring(l, 2)),
 			lua_tostring(l, 3),
 			lua_tostring(l, 4),
-			(ArgCount > 3) ? lua_tointeger(l, 5) : CPropUIControl::Priority_Default));
+			(ArgCount > 3) ? (int)lua_tointeger(l, 5) : CPropUIControl::Priority_Default));
 	return 1;
 }
 //---------------------------------------------------------------------
